Add parameter for granularity
There is a problem in the current commit and that it'll show slots as
available of the latter part of the hour is booked. I.e. if reality
looks like this:

17.00: busy
17.30: busy
18.00: free
18.30: busy
19.00: busy

Then the script will still show 18.00 as a free slot. For 30 minute
granularity that is correct, but not for the hourly granularity. Need to
fix this later on. For full empty hours things are working as they
should. Takeaway is, don't use the full hour yet!

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>